BrowserStack-driver: Added yardoc comments and improved testing
