+++ Rakefile
@@ -1,5 +1,6 @@
 require "bundler/gem_tasks"
 require "rake/testtask"
+require "yard"
 
 namespace :test do
   Rake::TestTask.new(:integration) do |t|
@@ -14,6 +15,10 @@ namespace :test do
   end
   desc "Run all tests"
   task :all => [:unit, :integration]
+end
+
+YARD::Rake::YardocTask.new do |t|
+  t.files   = ['lib/features/**/*']
 end
 
 task :default => ["test:unit"]
