+++ test/test_helper.rb
@@ -1,1 +1,2 @@
 require 'test/unit'
+require 'minitest/mock'
